+++ tests/Integration/Commands/QueuedJobCommand.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Cronboard\Tests\Integration\Commands;
+
+class QueuedJobCommand extends JobCommand
+{
+    
+}