+++ tests/TestCase.php
@@ -3,6 +3,7 @@
 namespace Cronboard\Tests;
 
 use Cronboard\CronboardServiceProvider;
+use Orchestra\Database\ConsoleServiceProvider;
 use Orchestra\Testbench\TestCase as OrchestraTestCase;
 
 abstract class TestCase extends OrchestraTestCase
@@ -11,7 +12,8 @@ abstract class TestCase extends OrchestraTestCase
 	{
 	    return [
 	    	CronboardServiceProvider::class,
-	    	CronboardTestsServiceProvider::class
+	    	CronboardTestsServiceProvider::class,
+	    	ConsoleServiceProvider::class,
 	    ];
 	}
 }